% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fade.R
\name{apply_fade_in}
\alias{apply_fade_in}
\title{Apply fade-in}
\usage{
apply_fade_in(wave, fade)
}
\arguments{
\item{wave}{A numeric vector longer than fade as made by \code{\link{make_wave}} for example}

\item{fade}{A numeric vector shorter than wave as made by \code{\link{make_fade}} for example}
}
\description{
Function used to apply a fade in effect to a wave
}
\seealso{
\code{\link{apply_fade_out}} \code{\link{apply_fade_in_out}}
}
