% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fade.R
\name{apply_fade_in_out}
\alias{apply_fade_in_out}
\title{Apply fade-out and in}
\usage{
apply_fade_in_out(wave, fade_in, fade_out = rev(fade_in))
}
\arguments{
\item{wave}{A numeric vector longer than fade as made by \code{\link{make_wave}} for example}

\item{fade_in}{A numeric vector shorter than wave as made by \code{\link{make_fade}} for example}

\item{fade_out}{A numeric vector shorter than wave as made by \code{\link{make_fade}} for example}
}
\description{
Function used to apply a fade out and in effect to a wave
}
\seealso{
\code{\link{apply_fade_in}} \code{\link{apply_fade_out}}
}
