% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fade.R
\name{apply_fade_out}
\alias{apply_fade_out}
\title{Apply fade-out}
\usage{
apply_fade_out(wave, fade)
}
\arguments{
\item{wave}{A numeric vector longer than fade as made by \code{\link{make_wave}} for example}

\item{fade}{A numeric vector shorter than wave as made by \code{\link{make_fade}} for example}
}
\description{
Function used to apply a fade out effect to a wave
}
\seealso{
\code{\link{apply_fade_in}} \code{\link{apply_fade_in_out}}
}
