% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fade.R
\name{make_fade}
\alias{make_fade}
\title{Make fade}
\usage{
make_fade(duration, sample = 44100)
}
\arguments{
\item{duration}{A float,double or int giving the duration of the fade in second}

\item{sample}{An int representing the sample rate in Hz. Standard sample rates are 11 025, 22 050, 44 100, 48 000 and 96 000}
}
\description{
A function to compute nice-ish fade in as a vector of number between 0 and 1
}
\examples{
\dontrun{fade <- make_fade(1)}
\dontrun{fade_48k <- make_fade(0.75,48000)}
}
