% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wave.R
\name{make_wave}
\alias{make_wave}
\title{Make Wave}
\usage{
make_wave(note_list, durations, harmonics = c(1))
}
\arguments{
\item{note_list}{vector of string containing the list of notes that should get played,
should have the same length as \code{durations} .}

\item{durations}{numeric vector containing the list of the duration of each notes in second,
should have the same length as \code{note_list} .}

\item{harmonics}{numeric vector containing ratio between harmonics the rank is determine by the place of the number in the vector.
Harmonics vectors can be build using \code{\link{build_harmonic}} or be found in the data frame \code{harmonics} which come with this package.}
}
\description{
This function create waves, savable as a wave file with \code{\link{save.wave}} or playable with if possible depending on the audio package,
from a vector of string representing notes and a numeric vector representing how long are each notes
}
