% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sin.R
\name{sum_wave}
\alias{sum_wave}
\title{Sum wave}
\usage{
sum_wave(wave1, wave2, nb_intersect)
}
\arguments{
\item{wave1}{A numeric vector longer than fade as made by \code{\link{make_wave}} for example}

\item{wave2}{A numeric vector longer than fade as made by \code{\link{make_wave}} for example}

\item{nb_intersect}{A integer representing the number a of point that should be sum in both waves should be lower than \code{min(length(wave1),length(wave2))}}
}
\description{
Utility function used to partially sum up two waves
}
